requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-toolbelt==0.0.1
django-widget-tweaks==1.4.5
gunicorn==20.0.4
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.2
selenium==3.141.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
whitenoise==5.0.1
